chore: refactor a pile of DTO code
