+++ sticker-award/src/main/java/com/datadoghq/stickerlandia/stickeraward/award/dto/UserAssignmentDTO.java
@@ -12,7 +12,7 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
     "assignedAt",
     "reason"
 })
-public class UserAssignment {
+public class UserAssignmentDTO {
 
     @JsonProperty("userId")
     private String userId;
